Change typedef for fpos_t

+++ libyaul/lib/lib/bits/alltypes.h
@@ -1,11 +1,11 @@
 typedef int pid_t;
-typedef unsigned id_t;
-typedef unsigned uid_t;
-typedef unsigned gid_t;
+typedef unsigned int id_t;
+typedef unsigned int uid_t;
+typedef unsigned int gid_t;
 typedef int key_t;
-typedef unsigned useconds_t;
-typedef long long off_t;
+typedef unsigned int useconds_t;
+typedef long off_t;
 typedef int ssize_t;
-typedef unsigned mode_t;
+typedef unsigned int mode_t;
 
 typedef struct _IO_FILE FILE;
